clouddrive/backend/requirements.txt
# KNOWN: CVE-2026-27205 (Missing Vary:Cookie header) - fix: upgrade to flask==3.1.3
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
Flask-Bcrypt==1.0.1
Flask-Limiter==3.8.0
PyJWT==2.12.0
psycopg2-binary==2.9.9
boto3==1.34.84
# KNOWN: CVE-2026-28684 (Symlink path traversal) - fix: upgrade to 1.2.2
python-dotenv==1.0.1
Werkzeug==3.1.6
# KNOWN: CVE-2024-1135, CVE-2024-6827 (HTTP Request Smuggling) - fix: upgrade to 22.0.0
gunicorn==21.2.0
Flask-Mail==0.10.0
